+++ paddle.hpp
@@ -25,7 +25,7 @@ class paddle_data : public unit
 public:
     paddle_data();
 
-    double get_x() const override;
+    int get_x() const override;
     int get_width() const override;
 
     void set_shrunken(bool shrunken);
