+++ unit.hpp
@@ -3,8 +3,8 @@
 
 struct coordinates
 {
-    double x;
-    double y;
+    int x;
+    int y;
 };
 
 class unit
@@ -15,8 +15,8 @@ class unit
 
 protected:
     void set_pos(const coordinates &pos);
-    void set_x(const double &x);
-    void set_y(const double &y);
+    void set_x(const int &x);
+    void set_y(const int &y);
     void set_width(const int &width);
     void set_height(const int &height);
 
@@ -24,13 +24,13 @@ public:
     virtual int get_width() const;
     virtual int get_height() const;
     virtual coordinates get_pos() const;
-    virtual double get_x() const;
-    virtual double get_y() const;
+    virtual int get_x() const;
+    virtual int get_y() const;
 
-    virtual double get_left() const;
-    virtual double get_right() const;
-    virtual double get_top() const;
-    virtual double get_bottom() const;
+    virtual int get_left() const;
+    virtual int get_right() const;
+    virtual int get_top() const;
+    virtual int get_bottom() const;
     virtual coordinates get_middle() const;
 };
